+++ src/gui/tasks.rs
@@ -333,6 +333,7 @@ pub fn import_cemu_settings(core: &Manager, path: &Path) -> Result<Message> {
             let deploy_config = wiiu_config.deploy_config.get_or_insert_default();
             deploy_config.auto = true;
             deploy_config.output = gfx_folder.join("BreathOfTheWild_UKMM");
+            deploy_config.executable = gfx_folder.with_file_name("Cemu.exe").exists_then();
         }
     } else {
         settings.wiiu_config = Some(PlatformSettings {
@@ -345,6 +346,7 @@ pub fn import_cemu_settings(core: &Manager, path: &Path) -> Result<Message> {
                     method: uk_manager::settings::DeployMethod::Copy,
                     output: gfx_folder.join("BreathOfTheWild_UKMM"),
                     cemu_rules: true,
+                    executable: gfx_folder.with_file_name("Cemu.exe").exists_then(),
                 }
             }),
         })